Definitely using a Multi-Key Hash Map using FNV 1a. Fixed more stuff, including Makefiles.

+++ lib/strings.h
@@ -1,4 +1,4 @@
 char* strfun(const char* input, int chartype (int));
-char** tokenize(const char* input, char** buffer, int chartype(int), const unsigned long len);
-char** spacetokenize(const char* input, char** buffer, const unsigned long len);
+char** tokenize(const char* input, char** buffer, int chartype(int), const unsigned int len);
+char** spacetokenize(const char* input, char** buffer, const unsigned int len);
 int isnewline(int character);
